requirements.txt
# Google API
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# Optional: load credentials from .env
python-dotenv>=1.0.0
